Implementar operaciones CRUD para Pedidos y refactorizar la estructura del modelo

+++ src/main/GridAndFlex/src/main/java/org/generation/GridAndFlex/repository/PedidosRepository.java
@@ -3,5 +3,5 @@ package org.generation.GridAndFlex.repository;
 import org.generation.GridAndFlex.model.Pedidos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidosRepository extends JpaRepository<Pedidos, Integer> {
+public interface PedidosRepository extends JpaRepository<Pedidos, Long> {
 }
